+++ CHANGELOG.md
@@ -16,6 +16,7 @@ accidentally triggering the load of a previous DB version.**
 * #4517 Fix prepared statement param handling in ChunkAppend
 * #4526 Fix gapfill group comparison for TOASTed values
 * #4527 Handle stats properly for range types
+* #4522 Fix ANALYZE on dist hypertable for a set of nodes
 
 **Thanks**